+++ src/file/file-runner.ts
@@ -8,6 +8,7 @@ import { SqlDialectAdapter } from "../models/SqlDialectAdapter";
 import postgres from '../sql-dialects/postgres';
 import duckdb from '../sql-dialects/duckdb';
 import oracle from '../sql-dialects/oracle';
+import sqlite from '../sql-dialects/sqlite';
 
 import { PitonFilePartResult } from "../models/PitonFilePartResult";
 import { PitonFileResult } from "../models/PitonFileResult";
@@ -41,6 +42,9 @@ export async function runFile(file: PitonFile): Promise<PitonFileResult> {
     }
     else if (file.sqlDialect === 'oracledb') {
         sql = oracle;
+    }
+    else if (file.sqlDialect === 'sqlite') {
+        sql = sqlite;
     }
 
     try {